Fix async queue to propagate cancellation all they way to the queue
