+++ exporter/exporterqueue/memory_queue_test.go
@@ -6,13 +6,10 @@ package exporterqueue
 import (
 	"context"
 	"sync"
-	"sync/atomic"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
-	"go.opentelemetry.io/collector/component/componenttest"
 )
 
 type sizerInt64 struct{}
@@ -107,62 +104,6 @@ func TestMemoryQueueZeroCapacity(t *testing.T) {
 	q := newMemoryQueue[int64](memoryQueueSettings[int64]{sizer: sizerInt64{}, capacity: 0})
 	require.ErrorIs(t, q.Offer(context.Background(), 1), ErrQueueIsFull)
 	require.NoError(t, q.Shutdown(context.Background()))
-}
-
-func TestAsyncMemoryQueue(t *testing.T) {
-	q := newMemoryQueue[int64](memoryQueueSettings[int64]{sizer: sizerInt64{}, capacity: 100})
-	consumed := &atomic.Int64{}
-	ac := newAsyncQueue(q, 1, func(_ context.Context, _ int64, done Done) {
-		consumed.Add(1)
-		done.OnDone(nil)
-	})
-	require.NoError(t, ac.Start(context.Background(), componenttest.NewNopHost()))
-	for j := 0; j < 10; j++ {
-		require.NoError(t, q.Offer(context.Background(), 10))
-	}
-	require.NoError(t, ac.Shutdown(context.Background()))
-	assert.EqualValues(t, 10, consumed.Load())
-}
-
-func TestAsyncMemoryQueueBlocking(t *testing.T) {
-	q := newMemoryQueue[int64](memoryQueueSettings[int64]{sizer: sizerInt64{}, capacity: 100, blocking: true})
-	consumed := &atomic.Int64{}
-	ac := newAsyncQueue(q, 4, func(_ context.Context, _ int64, done Done) {
-		consumed.Add(1)
-		done.OnDone(nil)
-	})
-	require.NoError(t, ac.Start(context.Background(), componenttest.NewNopHost()))
-	wg := &sync.WaitGroup{}
-	for i := 0; i < 10; i++ {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			for j := 0; j < 100_000; j++ {
-				assert.NoError(t, q.Offer(context.Background(), 10))
-			}
-		}()
-	}
-	wg.Wait()
-	require.NoError(t, ac.Shutdown(context.Background()))
-	assert.EqualValues(t, 1_000_000, consumed.Load())
-}
-
-func BenchmarkAsyncMemoryQueue(b *testing.B) {
-	q := newMemoryQueue[int64](memoryQueueSettings[int64]{sizer: sizerInt64{}, capacity: int64(10 * b.N)})
-	consumed := &atomic.Int64{}
-	require.NoError(b, q.Start(context.Background(), componenttest.NewNopHost()))
-	ac := newAsyncQueue(q, 1, func(_ context.Context, _ int64, done Done) {
-		consumed.Add(1)
-		done.OnDone(nil)
-	})
-	require.NoError(b, ac.Start(context.Background(), componenttest.NewNopHost()))
-	b.ResetTimer()
-	b.ReportAllocs()
-	for j := 0; j < b.N; j++ {
-		require.NoError(b, q.Offer(context.Background(), 10))
-	}
-	require.NoError(b, ac.Shutdown(context.Background()))
-	assert.EqualValues(b, b.N, consumed.Load())
 }
 
 func consume[T any](q readableQueue[T], consumeFunc func(context.Context, T) error) bool {
